chore(deps): upgrade wasmtime 18 → 44, eliminate 16 CVEs
Upgrades the optional `runtime-wasm` feature from wasmtime 18 to 44.
All 16 security advisories (RUSTSEC-2024-0438 through RUSTSEC-2026-0096)
are patched in wasmtime ≥ 43.0.1 and are removed from deny.toml.

## Changes
- `fraiseql-functions`: bump wasmtime dep to version 44
- `runtime/wasm/mod.rs`: adapt to wasmtime 44 host trait API changes
  - strip `wasmtime::Result<T>` wrapper from all host trait method returns
  - `instantiate` now returns single value (not tuple)
  - `add_to_linker` requires explicit `HasSelf<StoreData>` type parameter
- `runtime/wasm/store.rs`: update helper method signatures to match new
  host trait return types; add `# Errors` doc; make `get_env_var_value` const
- `runtime/wasm/limiter.rs`: `table_growing` params u32 → usize (wasmtime 44)
- `triggers/mutation.rs`: fix unused import, unnecessary raw string hashes,
  unwrap() → expect() in deno tests
- `routes/functions/mod.rs`: `&str` → `&'static str` (clippy pedantic)
- `deny.toml`: remove 16 stale wasmtime 18 CVE suppression entries;
  update bans.skip for wasmtime 44 duplicate dependency versions;
  remove ~10 other stale skip entries no longer in the dependency graph

## Verification
✅ cargo deny check — advisories ok, bans ok, licenses ok, sources ok
✅ cargo clippy -p fraiseql-functions --all-targets --all-features — clean
✅ cargo clippy -p fraiseql-server --lib --all-features — clean
✅ fraiseql-server lib tests: 953/953 pass
✅ fraiseql-functions lib tests: 163/163 pass (2 testcontainers skipped — Docker veth issue in this env, pre-existing)

+++ crates/fraiseql-functions/src/runtime/wasm/limiter.rs
@@ -125,9 +125,9 @@ impl ResourceLimiter for FunctionStoreLimiter {
     /// For now, we allow table growth without limits (only memory is limited).
     fn table_growing(
         &mut self,
-        _current: u32,
-        _desired: u32,
-        _maximum: Option<u32>,
+        _current: usize,
+        _desired: usize,
+        _maximum: Option<usize>,
     ) -> wasmtime::Result<bool> {
         Ok(true)
     }
@@ -176,7 +176,7 @@ mod tests {
         let mut limiter = FunctionStoreLimiter::new(1024 * 1024);
 
         // Table growth is always allowed
-        assert!(limiter.table_growing(0, 100, None).is_ok());
-        assert!(limiter.table_growing(100, 1000, None).is_ok());
+        assert!(limiter.table_growing(0_usize, 100_usize, None).is_ok());
+        assert!(limiter.table_growing(100_usize, 1000_usize, None).is_ok());
     }
 }
